Revamp signal and action system
The signalling system has been revamped to fit nicely with Qt models as
well as a thin web client in the future. All signal calls are virtual
methods on the tournament object and can be implemented in subclasses to
support relevant front ends.

All action implementations have been changed to perform as few signal
calls as possible and maintain a consistant and sane state despite being
given invalid parameters.

* Implement entire new signal system
* Reimplement actions with proper checks to maintain valid state
* Remove range checks from stores
* Add inclusion check methods to stores
* Consistently 'Erase' instead of delete/remove